corrected null pointer with PSC group
