+++ usaspending_api/references/tests/integration/filter_tree/psc/test_psc_basic_cases.py
@@ -73,21 +73,6 @@ def test_count_branching_at_middle(client, product_with_branching_count_above_on
     assert resp.json()["results"][0]["count"] == 2
 
 
-def test_wrong_path_at_tier_one(client, basic_service):
-    resp = _call_and_expect_200(client, base_query + "Product/B/")
-    assert resp.json() == {"results": []}
-
-
-def test_wrong_path_at_tier_two(client, basic_service):
-    resp = _call_and_expect_200(client, base_query + "Service/A/")
-    assert resp.json() == {"results": []}
-
-
-def test_wrong_path_at_tier_three(client, basic_service):
-    resp = _call_and_expect_200(client, base_query + "Service/C/B5/")
-    assert resp.json() == {"results": []}
-
-
 def _call_and_expect_200(client, url):
     resp = client.get(url)
     assert resp.status_code == status.HTTP_200_OK, "Failed to return 200 Response"
